Separate settings to their own module
