+++ src/App.tsx
@@ -26,7 +26,8 @@ import CreateWalletPages from './pages/WalletManagement/CreateWalletRootPage'
 import ImportWalletPages from './pages/WalletManagement/ImportWalletRootPage'
 import WalletPages from './pages/Wallet/WalletRootPage'
 import SettingsPage from './pages/SettingsPage'
-import { createClient, loadSettings, saveSettings, Settings } from './utils/clients'
+import { createClient } from './utils/clients'
+import { loadSettings, saveSettings, Settings } from './utils/settings'
 import { useInterval, useStateWithLocalStorage } from './utils/hooks'
 import { Modal } from './components/Modal'
 import Spinner from './components/Spinner'
