+++ src/components/Address.tsx
@@ -17,7 +17,7 @@
 import styled from 'styled-components'
 import tinycolor from 'tinycolor2'
 
-import { loadSettings } from '../utils/clients'
+import { loadSettings } from '../utils/settings'
 import { openInWebBrowser } from '../utils/misc'
 
 const Address = ({ hash }: { hash: string }) => {
