+++ src/pages/Wallet/WalletHomePage.tsx
@@ -33,7 +33,7 @@ import AppHeader, { HeaderDivider } from '../../components/AppHeader'
 import Address from '../../components/Address'
 import NetworkBadge from '../../components/NetworkBadge'
 import { abbreviateAmount, calAmountDelta } from '../../utils/numbers'
-import { loadSettings, useCurrentNetwork } from '../../utils/clients'
+import { loadSettings, useCurrentNetwork } from '../../utils/settings'
 import { useInterval } from '../../utils/hooks'
 import { getHumanReadableError } from '../../utils/api'
 import { openInWebBrowser } from '../../utils/misc'
